requirements.txt
fastapi
uvicorn
pyodbc
azure-identity
azure-keyvault-secrets
azure.storage.blob
fastavro
